ui fixes
