+++ src/app.html
@@ -115,7 +115,7 @@
 
 	<body data-sveltekit-preload-data="hover" class="bg-gray-100">
 		<!-- Header -->
-		<header class="bg-white border-b sticky top-0 z-10">
+		<!-- <header class="bg-white border-b sticky top-0 z-10">
 			<div class="max-w-6xl mx-auto px-4 flex items-center justify-between h-16">
 				<div class="flex items-center">
 					<div class="text-blue-700 font-bold text-3xl mr-8">LinkedLens</div>
@@ -129,7 +129,7 @@
 				</div>
 
 				<nav class="flex items-center space-x-6">
-					<a href="/feed" class="flex flex-col items-center text-gray-500 hover:text-black">
+					<a href="#" class="flex flex-col items-center text-gray-500 hover:text-black">
 						<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="text-blue-700">
 							<path d="m3 9 9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z"/>
 							<polyline points="9 22 9 12 15 12 15 22"/>
@@ -152,7 +152,7 @@
 						</svg>
 						<span class="text-xs mt-1">Jobs</span>
 					</a>
-					<a href="/" class="flex flex-col items-center text-gray-500 hover:text-black relative">
+					<a href="/lens" class="flex flex-col items-center text-gray-500 hover:text-black relative">
 						<svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="text-blue-700">
 							<path d="M12 2a9 9 0 1 0 9 9"/>
 							<path d="M12 2a4 4 0 0 0-4 4"/>
@@ -168,7 +168,7 @@
 					<div class="h-8 w-8 rounded-full bg-gray-300 cursor-pointer"></div>
 				</nav>
 			</div>
-		</header>
+		</header> -->
 
 		<div style="display: contents">%sveltekit.body%</div>
