+++ src/lib/components/chat/Navbar.svelte
@@ -27,7 +27,7 @@
 
 	import PencilSquare from '../icons/PencilSquare.svelte';
 	import Home from '../icons/Home.svelte';
-	import {goto} from "$app/navigation";
+	import { goto } from '$app/navigation';
 
 	const i18n = getContext('i18n');
 
@@ -42,7 +42,7 @@
 	let showShareChatModal = false;
 	let showDownloadChatModal = false;
 	const goToFeed = async () => {
-		await goto('/feed');
+		await goto('/');
 	};
 </script>
 
@@ -121,24 +121,7 @@
 					</Menu>
 				{/if}
 
-				<Tooltip content={$i18n.t('Feed')}>
-					<button
-							id="feed-button"
-							class=" flex {$showSidebar
-							? 'md:hidden'
-							: ''} cursor-pointer px-2 py-2 rounded-xl text-gray-600 dark:text-gray-400 hover:bg-gray-50 dark:hover:bg-gray-850 transition"
-							on:click={() => {
-							goToFeed();
-						}}
-							aria-label="New Chat"
-					>
-						<div class=" m-auto self-center">
-							<Home className=" size-5" strokeWidth="2" />
-						</div>
-					</button>
-				</Tooltip>
-
-				<Tooltip content={$i18n.t('Controls')}>
+				<!-- <Tooltip content={$i18n.t('Controls')}>
 					<button
 						class=" flex cursor-pointer px-2 py-2 rounded-xl hover:bg-gray-50 dark:hover:bg-gray-850 transition"
 						on:click={async () => {
@@ -150,9 +133,24 @@
 							<AdjustmentsHorizontal className=" size-5" strokeWidth="0.5" />
 						</div>
 					</button>
+				</Tooltip> -->
+
+				<Tooltip content={$i18n.t('Feed')}>
+					<button
+						id="feed-button"
+						class=" flex cursor-pointer px-2 py-2 rounded-xl text-gray-600 dark:text-gray-400 hover:bg-gray-50 dark:hover:bg-gray-850 transition"
+						on:click={() => {
+							goToFeed();
+						}}
+						aria-label="Feed"
+					>
+						<div class=" m-auto self-center">
+							<Home className=" size-5" strokeWidth="2" />
+						</div>
+					</button>
 				</Tooltip>
 
-				<Tooltip content={$i18n.t('New Chat')}>
+				<!-- <Tooltip content={$i18n.t('New Chat')}>
 					<button
 						id="new-chat-button"
 						class=" flex {$showSidebar
@@ -167,7 +165,7 @@
 							<PencilSquare className=" size-5" strokeWidth="2" />
 						</div>
 					</button>
-				</Tooltip>
+				</Tooltip> -->
 
 				{#if $user !== undefined && $user !== null}
 					<UserMenu
