+++ src/routes/lens/+page.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+	import Chat from '$lib/components/chat/Chat.svelte';
+	import Help from '$lib/components/layout/Help.svelte';
+</script>
+
+<Help />
+<Chat />
